getCoinPrices moved to wallet object
